% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap}
\usage{
myboot2(iter, x, fun, alpha, cx, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{the vector of data}

\item{fun}{the function to be used (e.g. "mean")}

\item{alpha}{alpha value; for the confidence interval}

\item{cx}{font size of histogram labels}

\item{...}{other formatting options}
}
\value{

}
\description{
Runs a bootstrap analysis of the given data based on the function you would like to run.
Provides a histogram of the bootstrap analysis with confidence interval labels
}
\examples{
myboot2(iter=10000,x,fun="mean",alpha=0.05,cx=1.5,...)
With 10000 iterations of the data stored in x with the function
performing a bootstrap for the mean of the data stored in x.
This example uses an alpha level of 0.05, so it has a 95\%
confidence interval. It has a font size setting of 1.5.
}
