% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit Theorem Uniform Distribution}
\usage{
mycltu(n, iter, a, b)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{a}{x-axis minimum value}

\item{b}{x-axis maximum value}
}
\value{
A histogram of a uniform distribution following the central limit theorem
}
\description{
Using this function will create a uniform distribution of
the data as provided. Giving a sample size, number of iterations,
and axis minimum and maximum values, the function will
create a histogram with a normal curve and a uniform curve.
}
\examples{
mycltu(n=30,iter=10000,a=0,b=10)
Where n = 30 sample size, iter = 10000 iterations of the sample,
a = 0 minimum value, and b = 10 maximum value.
}
