% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Maximum Likelihood Beta}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{vector of data}

\item{alpha}{the alpha vector}

\item{beta}{the beta vector}

\item{...}{can insert other information for formatting the beta distribution}
}
\value{
a beta distribution from the inputted data
}
\description{
Maximum Likelihood Beta
}
\examples{
mymlbeta(x=rbeta(100,shape1=2,shape2=5),alpha=seq(1,3,length=100),beta=seq(2,8,length=100),lwd=2,labcex=1)
}
