% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbinom.R
\name{mynbin}
\alias{mynbin}
\title{Negative Binomial Probability}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{number of trials}

\item{r}{number of successes}

\item{p}{hypothesized probability of successes}
}
\value{

}
\description{
This function will provide the probability of a certain number
of successes of binomial trials when given y, the number of trials,
r, the number of successes, and p, the probability of a success.
The probability is provided as a decimal, which can be understood
as a percent once multiplied by 100.
}
\examples{
mynbin(10,3,0.4)
Where y = 10 trials, r = 3 successes, and 0.4 is the probability of a success
}
