% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean value}

\item{sigma}{standard deviation value}

\item{a}{the value that alpha cuts off at; for P(x<=a)}
}
\value{

}
\description{
This function wil produce a normal curve when given a mean value,
a standard deviation, and a significance level point estimate
where alpha cuts off.
This function will provide a shaded region to the left of the
provided cut off value (a).
This function also provides the probability of a value at or
below the given 'a' value.
}
\examples{
myncurve(10,5,6)
With mu = 10 mean, sigma = 5 standard deviation, and a = 6
where the alpha level cut-off happens.
So this will be a normal distirbution centered at 10 with a
tail shaded in to the left of the value 6.
There will also be a provided probability of a value at or below
'a', so the probability of a value at or below 6.
}
